Egad! It compiles!
I don't know what I was thinking, but after 10+ years of programming, I should know better than to try to make enums and consts static. Embarrassments aside, all constant settings have been moved to a settings class so they are in one place and easily modifiable. The state machine has been expanded upon a bit and is now operational. Future todo's are to add transitions between states so it's not so abrupt, but that's aesthetics more than anything.